diet_composer start

+++ apps/users/models/models.py
@@ -21,8 +21,10 @@ class Profile(models.Model):
     image = models.ImageField(default="default.jpg", upload_to="profile_pics")
     age = models.PositiveSmallIntegerField(validators=[validate_age], null=True, blank=True)
     gender = models.CharField(choices=Gender.choices, max_length=10,  null=True, blank=True)
-    height = models.FloatField(null=True, blank=True, help_text="height in centimeters")
-    weight = models.FloatField(null=True, blank=True, help_text="weight in kilograms")
+    height = models.DecimalField(max_digits=4, decimal_places=1,
+                                 null=True, blank=True, help_text="height in centimeters")
+    weight = models.DecimalField(max_digits=4, decimal_places=1,
+                                 null=True, blank=True, help_text="weight in kilograms")
     activity = models.ForeignKey(UserActivity, on_delete=models.PROTECT, null=True, blank=True)
 
     def __repr__(self) -> str:
